Rename folder that contains dummy business logic
