+++ examples/bri-3/src/bri/identity/bpiSubjects/capabilities/createBpiSubject/createBpiSubject.command.ts
