+++ examples/bri-3/src/bri/identity/bpiSubjects/controllers/subjects.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { CommandBus } from '@nestjs/cqrs';
-import { CreateBpiSubjectCommand } from '../businessLogic/createBpiSubject/createBpiSubject.command';
+import { CreateBpiSubjectCommand } from '../capabilities/createBpiSubject/createBpiSubject.command';
 
 @Controller("subjects")
 export class SubjectController {
