+++ examples/bri-3/src/bri/identity/bpiSubjects/subjects.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { SubjectController } from './controllers/subjects.controller';
-import { CreateBpiSubjectCommandHandler } from './businessLogic/createBpiSubject/createBpiSubjectCommand.handler';
+import { CreateBpiSubjectCommandHandler } from './capabilities/createBpiSubject/createBpiSubjectCommand.handler';
 
 export const CommandHandlers = [CreateBpiSubjectCommandHandler];
